@mossr 2023-12-15
I'm not a fan of the `$` interpolation as a requirement. Could you explain that choice?