@zsunberg 2023-12-15
> I'm not a fan of the `$` interpolation as a requirement. Could you explain that choice?
That's just for benchmarking: https://juliaci.github.io/BenchmarkTools.jl/stable/manual/#Interpolating-values-into-benchmark-expressions