@mossr 2023-12-15
Ah that's great. So I'm guessing it will infer the type based on `x` here? (Instead of my implementation with an explicit type input)