@zsunberg 2023-12-15
> Ah that's great. So I'm guessing it will infer the type based on x here? (Instead of my implementation with an explicit type input)

Exactly.

One downside is that this would stop working for `RectangleGrid` interpolations in more than 16 dimensions:
```
julia> @MVector(zeros(Int, 2^17))
Internal error: encountered unexpected error during compilation of zeros:
StackOverflowError()
```
Probably people should be using simplex interpolation by that point though...