@mossr 2023-12-15
I think that's reasonable and should just be mentioned in the README to warn users (and offer a suggestion to use the `SimplexGrid` instead)