@mossr 2023-12-16
I pulled and tested this locally and I get the following errors when trying `Complex` or `Real` typed `x` vectors:
```julia
grid = RectangleGrid([0.0, 0.5, 1.0], [0.0, 0.5, 1.0])
grid_data = [8.0, 1.0, 6.0, 3.0, 5.0, 7.0, 4.0, 9.0, 2.0]
x = Complex[0.25, 0.75]
interpolate(grid, grid_data, x)

ERROR: setindex!() with non-isbitstype eltype is not supported by StaticArrays. Consider using SizedArray.
Stacktrace:
 [1] error(s::String)
   @ Base .\error.jl:35
 [2] setindex!
   @ C:\Users\RobertMoss\.julia\packages\StaticArrays\cZ1ET\src\MArray.jl:39 [inlined]
 [3] interpolants(grid::RectangleGrid{2}, x::Vector{Complex})
   @ GridInterpolations C:\Users\RobertMoss\.julia\packages\GridInterpolations\CSHS7\src\GridInterpolations.jl:168
 [4] interpolate(grid::RectangleGrid{2}, data::Vector{Float64}, x::Vector{Complex})
   @ GridInterpolations C:\Users\RobertMoss\.julia\packages\GridInterpolations\CSHS7\src\GridInterpolations.jl:145
 [5] top-level scope
   @ REPL[17]:1
```

Do you see this as well?