@zsunberg 2023-12-16
Yeah, I realized that I had a typo above which meant that I had not actually tested the `Complex` version.

One problem with what you tried is that `Complex` is an abstract type. If you try `Complex{Float64}` you will get past that error. But then you will run into the error that you cannot compare complex with floating point. In the end, it just doesn't make sense to interpolate complex numbers onto a floating point grid.

Really this PR just makes things differentiable and (imo just as importantly) eliminates returning references to memory that will be silently re-written later. It doesn't add the capability to accept more exotic arguments (e.g. complex). That would require more work.

(I just edited and removed the `Complex` example from the PR description since it is not meaningful)