@mossr 2023-12-16
I'm good with that. The purpose of my PR #43 was solely to get autodiff working, and I tested the `ForwardDiff.gradient(f, x)` example and it works.

Looks good to me!